@@ -1,4 +1,4 @@
-class CreateProjectUpdates < ActiveRecord::Migration
+class CreateProjectUpdates < ActiveRecord::Migration[4.2]
   def change
     create_table :project_updates do |t|
       t.integer :project_id
