@@ -1,4 +1,4 @@
-class RemoveDeleteFlagFromProjects < ActiveRecord::Migration
+class RemoveDeleteFlagFromProjects < ActiveRecord::Migration[4.2]
   def change
     remove_column :projects, :delete_flag, :integer
   end
