@@ -1,4 +1,4 @@
-class CreateSkills < ActiveRecord::Migration
+class CreateSkills < ActiveRecord::Migration[4.2]
   def change
     create_table :skills do |t|
       t.string :name
