@@ -1,4 +1,4 @@
-class AddIndexNameToSkills < ActiveRecord::Migration
+class AddIndexNameToSkills < ActiveRecord::Migration[4.2]
   def change
       add_index :skills, [:name], :unique => true
   end
