@@ -1,4 +1,4 @@
-class RenameUsersSkillsToSkillsUsers < ActiveRecord::Migration
+class RenameUsersSkillsToSkillsUsers < ActiveRecord::Migration[4.2]
   def change
       rename_table :users_skills, :skills_users
   end
