@@ -1,4 +1,4 @@
-class CreateProjectsUsers < ActiveRecord::Migration
+class CreateProjectsUsers < ActiveRecord::Migration[4.2]
   def change
     create_table :projects_users do |t|
       t.integer :project_id
