@@ -1,4 +1,4 @@
-class AddColumnToUsers < ActiveRecord::Migration
+class AddColumnToUsers < ActiveRecord::Migration[4.2]
   def change
       add_column :users, :provider, :string
       add_column :users, :uid, :string
