@@ -1,4 +1,4 @@
-class RemoveColumnUser < ActiveRecord::Migration
+class RemoveColumnUser < ActiveRecord::Migration[4.2]
   def change
       remove_column :users, :facebook_user_id
       remove_column :users, :password
