@@ -1,4 +1,4 @@
-class DropTableUserSession < ActiveRecord::Migration
+class DropTableUserSession < ActiveRecord::Migration[4.2]
   def down
       drop_table :user_sessions
   end
