@@ -1,4 +1,4 @@
-class DeleteIndexUser < ActiveRecord::Migration
+class DeleteIndexUser < ActiveRecord::Migration[4.2]
   def change
       remove_index :users, :email
   end
