@@ -1,4 +1,4 @@
-class CreateStages < ActiveRecord::Migration
+class CreateStages < ActiveRecord::Migration[4.2]
   def change
     create_table :stages do |t|
       t.string :name
