@@ -1,4 +1,4 @@
-class ChangeColumn2ToProject < ActiveRecord::Migration
+class ChangeColumn2ToProject < ActiveRecord::Migration[4.2]
   def change
     change_column :projects, :description, :text
   end
