@@ -1,4 +1,4 @@
-class ChangeColumnToProjectUpdate < ActiveRecord::Migration
+class ChangeColumnToProjectUpdate < ActiveRecord::Migration[4.2]
   def change
     change_column :project_updates, :description, :text
   end
