@@ -1,5 +1,5 @@
 # coding: utf-8
-class ChangeColumnToProject2 < ActiveRecord::Migration
+class ChangeColumnToProject2 < ActiveRecord::Migration[4.2]
   def up
     change_column :projects, :stage_id, :integer, default: 10
   end
