@@ -1,4 +1,4 @@
-class AddProjectToProjects < ActiveRecord::Migration
+class AddProjectToProjects < ActiveRecord::Migration[4.2]
   def change
     add_reference :projects, :project, index: true, foreign_key: true
   end
