@@ -1,4 +1,4 @@
-class CreateProjectUpdateHistories < ActiveRecord::Migration
+class CreateProjectUpdateHistories < ActiveRecord::Migration[4.2]
   def change
     create_table :project_update_histories do |t|
       t.references :project_update, index: true, foreign_key: true
