@@ -1,4 +1,4 @@
-class AddReceiveAllToUser < ActiveRecord::Migration
+class AddReceiveAllToUser < ActiveRecord::Migration[4.2]
   def change
     add_column :users, :receive_all, :bool
   end
